requirements.txt
django-stubs==4.2.0
djangorestframework-stubs==3.14.0
types-python-dateutil==2.8.19
types-requests==2.31.0.10
